Move handshake logic to module
