+++ crates/networking/p2p_actors/src/peer/mod.rs
@@ -3,6 +3,7 @@ pub mod ingress;
 
 pub mod constants;
 mod crypto;
+mod handshake;
 pub mod utils;
 
 pub use actor::{Actor, Config, Error};
